% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ERA_13_18_bib_cleaned}
\alias{ERA_13_18_bib_cleaned}
\title{ERA CSA search bibliography 2013-2018}
\format{
A data frame with rows representing research articles and the following columns:
\describe{
\item{\code{CATEGORY}}{Type of publication}
\item{\code{AUTHOR}}{Authors of the publication}
\item{\code{BOOKTITLE}}{Title of the book, NA if the publication is in a journal}
\item{\code{CHAPTER}}{The chapter number in the book}
\item{\code{EDITION}}{Edition of the book}
\item{\code{EDITOR}}{Editors of the publication}
\item{\code{INSTITUTION}}{Institution associated with the publication}
\item{\code{JOURNAL}}{Journal in which the article was published}
\item{\code{MONTH}}{Month of publication}
\item{\code{NUMBER}}{Journal number within the year}
\item{\code{ORGANIZATION}}{Organization associated with the publication}
\item{\code{PAGES}}{Page numbers of the publication}
\item{\code{PUBLISHER}}{Publisher of the publication}
\item{\code{SERIES}}{Series name of the publication}
\item{\code{TITLE}}{Title of the article}
\item{\code{TYPE}}{Type of article}
\item{\code{VOLUME}}{Volume of the journal}
\item{\code{YEAR}}{Year of publication}
\item{\code{DOI}}{Digital Object Identifier of the article}
\item{\code{ISSN}}{International Standard Serial Number of the journal}
\item{\code{URL}}{Web address of the journal article}
\item{\code{ERACODE}}{Unique data entry code}
\item{\code{KEYWORDS}}{Keywords used in the article}
\item{\code{ABSTRACT}}{Brief summary of the publication}
\item{\code{FILE}}{File path or identifier within the dataset}
\item{\code{ISBN}}{International Standard Book Number, applicable only for books}
\item{\code{NoAccess}}{Indicates if the paper could not be accessed due to subscription issues}
\item{\code{Passed.FT}}{Indicates if the article passed full text screening}
\item{\code{Passed.TA}}{Indicates if the article passed title and abstract screening}
}
}
\source{
WoS and Scopus ERA search results
}
\usage{
ERA_13_18_bib_cleaned
}
\description{
This dataset contains the results of deduplicated WoS and Scopus ERA search
for the years 2013-2018. The searches, detailed in \code{ERA_Search_Terms} from the ERAg package,
were conducted in May 2018. The search covers Africa.
}
\keyword{datasets}
